Removed sign-compare warnings.
libsofia-sip-ua/su/su_port.c
libsofia-sip-ua/su/su_wait.c
libsofia-sip-ua/iptsec/auth_client.c
utils/sip-date.c

darcs-hash:20061211160632-65a35-5ea2fa335ec43c85ef6a51f83a98bab02684df83.gz

+++ libsofia-sip-ua/iptsec/auth_client.c
@@ -256,7 +256,8 @@ int ca_info(auth_client_t *ca,
     return 0;
 
   if (!ca->ca_auc
-      || ca->ca_auc->auc_plugin_size <= offsetof(auth_client_plugin_t, auc_info)
+      || (size_t)ca->ca_auc->auc_plugin_size <= 
+         offsetof(auth_client_plugin_t, auc_info)
       || !ca->ca_auc->auc_info)
     return 0;
